Do not spend most of Travis-ci's time building a node runtime
Travis only needs to run tests which should work on a stock runtime.

+++ scripts/postinstall.sh
@@ -15,30 +15,35 @@ popd >& /dev/null
 
 [ -f "$PROJECT_ROOT"/package.json ]
 
-if ! [ -x "$PROJECT_ROOT/bin/node" ]; then
+# Don't bother building the runtime for Travis which can run tests on a stock runtime
+# docs.travis-ci.com/user/environment-variables/#default-environment-variables
+if [[ "true" != "$TRAVIS" ]]; then
 
-    NODE_BUILD_PARENT="$PROJECT_ROOT/bin/node.d"
-    NODE_BUILD_DIR="$NODE_BUILD_PARENT/node"
+    if ! [ -x "$PROJECT_ROOT/bin/node" ]; then
 
-    [ -d "$NODE_BUILD_PARENT" ]
-    rm -rf "$NODE_BUILD_DIR"
+        NODE_BUILD_PARENT="$PROJECT_ROOT/bin/node.d"
+        NODE_BUILD_DIR="$NODE_BUILD_PARENT/node"
 
-    (
-        pushd "$NODE_BUILD_PARENT"
-        git clone git@github.com:nodejs/node.git node
-        pushd "$NODE_BUILD_DIR"
-        git reset --hard edb03cb65d51e336713d6af1134a7394c0776635
-        patch -p1 < "$NODE_BUILD_PARENT/node.patch"
-        ./configure
-        make -j4
-        popd
-        popd
-    )
+        [ -d "$NODE_BUILD_PARENT" ]
+        rm -rf "$NODE_BUILD_DIR"
 
-    cp "$NODE_BUILD_DIR"/node "$PROJECT_ROOT"/bin/
+        (
+            pushd "$NODE_BUILD_PARENT"
+            git clone git@github.com:nodejs/node.git node
+            pushd "$NODE_BUILD_DIR"
+            git reset --hard edb03cb65d51e336713d6af1134a7394c0776635
+            patch -p1 < "$NODE_BUILD_PARENT/node.patch"
+            ./configure
+            make -j4
+            popd
+            popd
+        )
+
+        cp "$NODE_BUILD_DIR"/node "$PROJECT_ROOT"/bin/
+    fi
+
+    [ -x "$PROJECT_ROOT/bin/node" ]
+
+    # Sanity check our patched version of node.
+    [[ "true" == "$("$PROJECT_ROOT/bin/node" -e 'console.log(typeof require.keys.publicKey === `function`)')" ]]
 fi
-
-[ -x "$PROJECT_ROOT/bin/node" ]
-
-# Sanity check our patched version of node.
-[[ "true" == "$("$PROJECT_ROOT/bin/node" -e 'console.log(typeof require.keys.publicKey === `function`)')" ]]
